+++ pysubgroup/tests/test_algorithms_boolean.py
@@ -84,7 +84,7 @@ class TestAlgorithms(TestAlgorithmsBase, BooleanTargetBase, unittest.TestCase):
                           0.0452,
                           0.044399999999999995]
         data = get_credit_data()
-        target = ps.NominalTarget('class', b'bad')
+        target = ps.BinaryTarget('class', b'bad')
         searchSpace = ps.create_nominal_selectors(data, ignore=['class'])
         self.task = ps.SubgroupDiscoveryTask(data, target, searchSpace, result_set_size=10, depth=5, qf=ps.StandardQF(1.0))
 
@@ -126,7 +126,7 @@ class TestAlgorithms2(TestAlgorithmsBase, BooleanTargetBase, unittest.TestCase):
                           0.044399999999999995
                           ]
         data = get_credit_data()
-        target = ps.NominalTarget('class', b'bad')
+        target = ps.BinaryTarget('class', b'bad')
         searchSpace_Nominal = ps.create_nominal_selectors(data, ignore=['class'])
         searchSpace_Numeric = ps.create_numeric_selectors(data, ignore=['class'])
         searchSpace = searchSpace_Nominal + searchSpace_Numeric
@@ -179,7 +179,7 @@ class TestAlgorithms3(TestAlgorithmsBase, BooleanTargetBase, unittest.TestCase):
                         0.10813405094128754,
                         ]
         data = get_credit_data()
-        target = ps.NominalTarget('class', b'bad')
+        target = ps.BinaryTarget('class', b'bad')
         searchSpace_Nominal = ps.create_nominal_selectors(data, ignore=['class'])
         searchSpace_Numeric = ps.create_numeric_selectors(data, ignore=['class'])
         searchSpace = searchSpace_Nominal + searchSpace_Numeric
