+++ pysubgroup/tests/test_visualization.py
@@ -7,7 +7,7 @@ if __name__ == '__main__':
 
 
     data = get_titanic_data()
-    target = ps.NominalTarget('Survived', 0)
+    target = ps.BinaryTarget('Survived', 0)
     search_space = ps.create_selectors(data, ignore=['Survived'])
     task = ps.SubgroupDiscoveryTask(data, target, search_space,
                                     result_set_size=5, depth=2,
